bump

+++ src/testing/loadtest-azure/scripts/loadtest-get-files.ps1
@@ -12,18 +12,31 @@ param
   # Load Test data plane api version
   [string] $apiVersion = "2022-06-01-preview",
 
+  # optional - request an individual file via its fileId
+  [string] $fileId,
+
+  # optional - keep access token when used embedded
+  [bool] $keepToken = $false,
+
   [int] $maxPageSize
 )
 
 . "$PSScriptRoot/common.ps1"
 
-$urlRoot = "https://" + $apiEndpoint + "/loadtests/" + $loadTestId + "/files"
+$urlRoot = "https://{0}/loadtests/{1}/files" -f $apiEndpoint, $loadTestId
+
+if ($fileId) {
+  $urlRoot = "{0}/{1}" -f $urlRoot,$fileId
+}
 
 az rest --url $urlRoot `
   --method GET `
   --skip-authorization-header `
   --headers ('@' + $accessTokenFileName) `
   --url-parameters api-version=$apiVersion maxPageSize=$maxPageSize `
-  $verbose
+  $verbose --output json | convertFrom-Json
 
-Remove-Item $accessTokenFileName
+if (!$keepToken) {
+  # delete accessToken when $keepToken is set to $false (default)
+  Remove-Item $accessTokenFileName
+}
